+++ src/axom/mint/tests/mint_mesh_face_relation.cpp
@@ -8,8 +8,8 @@
 
 #include "axom/core/utilities/Utilities.hpp" /* for utilities::max */
 
-#include "axom/slic/core/UnitTestLogger.hpp" /* for UnitTestLogger */
-#include "axom/slic/interface/slic.hpp"      /* for slic macros */
+#include "axom/slic/core/SimpleLogger.hpp" /* for SimpleLogger */
+#include "axom/slic/interface/slic.hpp"    /* for slic macros */
 
 #include "gtest/gtest.h" /* for TEST and EXPECT_* macros */
 
@@ -1440,14 +1440,14 @@ TEST(mint_mesh_face_relation, correct_construction)
 }
 
 //------------------------------------------------------------------------------
-#include "axom/slic/core/UnitTestLogger.hpp"
-using axom::slic::UnitTestLogger;
+#include "axom/slic/core/SimpleLogger.hpp"
+using axom::slic::SimpleLogger;
 
 int main(int argc, char *argv[])
 {
   ::testing::InitGoogleTest(&argc, argv);
 
-  UnitTestLogger logger;  // create & initialize test logger,
+  SimpleLogger logger;  // create & initialize test logger,
   // finalized when exiting main scope
 
   return RUN_ALL_TESTS();