+++ src/axom/sidre/tests/sidre_mfem_datacollection.cpp
@@ -172,8 +172,8 @@ TEST(sidre_datacollection, dc_par_reload)
 }
 
     //----------------------------------------------------------------------
-    #include "axom/slic/core/UnitTestLogger.hpp"
-using axom::slic::UnitTestLogger;
+    #include "axom/slic/core/SimpleLogger.hpp"
+using axom::slic::SimpleLogger;
 
 int main(int argc, char* argv[])
 {
@@ -181,7 +181,7 @@ int main(int argc, char* argv[])
 
   ::testing::InitGoogleTest(&argc, argv);
 
-  UnitTestLogger logger;  // create & initialize test logger,
+  SimpleLogger logger;  // create & initialize test logger,
 
   MPI_Init(&argc, &argv);
   result = RUN_ALL_TESTS();