+++ src/axom/slam/tests/slam_set_RangeSet.cpp
@@ -419,7 +419,7 @@ int main(int argc, char* argv[])
 #endif
 
   // create & initialize test logger, finalized when exiting main scope
-  axom::slic::UnitTestLogger logger;
+  axom::slic::SimpleLogger logger;
   //axom::slic::debug::checksAreErrors = true;
 
   result = RUN_ALL_TESTS();